chore: extract row check from where clause
