+++ src/main.rs
@@ -14,6 +14,7 @@ mod database;
 mod row; // TODO: maybe put it inside database or table?
 mod query_result;
 mod where_clause;
+mod row_check;
 mod connection;
 mod execution_error;
 mod meta_command_error;
