+++ src/where_clause.rs
@@ -1,49 +1,7 @@
 use crate::lexer::SqlValue;
-use crate::table::error::TableError;
-use crate::row::Row;
 use crate::table::ColumnType;
 use crate::cmp_operator::CmpOperator;
-
-#[derive(Debug)]
-pub enum WhereValue {
-    TableColumn(usize),
-    Value(SqlValue),
-}
-
-#[derive(Debug)]
-pub struct WhereFilter<'a> {
-    operator: CmpOperator,
-    left: WhereValue,
-    right: WhereValue,
-    column_types: &'a [ColumnType],
-}
-
-impl<'a> WhereFilter<'a> {
-    pub fn dummy() -> Self {
-        Self {
-            operator: CmpOperator::Equals,
-            left: WhereValue::Value(SqlValue::Integer(1)),
-            right: WhereValue::Value(SqlValue::Integer(1)),
-            column_types: &[],
-        }
-    }
-
-    pub fn matches(&'a self, row: &'a Row) -> Result<bool, TableError> {
-        self
-            .operator
-            .apply(&self.get_value(&self.left, row)?, &self.get_value(&self.right, row)?)
-            .map_err(TableError::CmpError)
-
-    }
-
-    fn get_value(&'a self, value: &'a WhereValue, row: &'a Row) -> Result<SqlValue, TableError> {
-        match value {
-            WhereValue::Value(sql_value) => Ok(sql_value.clone()),
-            WhereValue::TableColumn(index) =>
-                row.get_cell_sql_value(self.column_types, *index).map_err(TableError::CannotGetCell),
-        }
-    }
-}
+use crate::row_check::{RowCheck, RowCheckValue};
 
 #[derive(Debug)]
 pub struct WhereClause {
@@ -53,11 +11,11 @@ pub struct WhereClause {
 }
 
 impl WhereClause {
-    pub fn compile<'a>(self, column_types: &'a [ColumnType], table_name: &'a str, column_names: &'a [String]) -> WhereFilter<'a> {
+    pub fn compile<'a>(self, column_types: &'a [ColumnType], table_name: &'a str, column_names: &'a [String]) -> RowCheck<'a> {
         let left = Self::build_where_value(self.left_value, table_name, column_names);
         let right = Self::build_where_value(self.right_value, table_name, column_names);
 
-        WhereFilter {
+        RowCheck {
             operator: self.operator,
             left,
             right,
@@ -65,7 +23,7 @@ impl WhereClause {
         }
     }
 
-    pub fn build_where_value(value: SqlValue, table_name: &str, column_names: &[String]) -> WhereValue {
+    pub fn build_where_value(value: SqlValue, table_name: &str, column_names: &[String]) -> RowCheckValue {
         let string_value = value.to_string();
         let splitted_identificator: Vec<&str> = string_value.split('.').collect();
         match splitted_identificator.len() {
@@ -73,23 +31,23 @@ impl WhereClause {
                 let index = column_names.iter()
                     .position(|table_column_name| table_column_name.eq(&string_value));
                 match index {
-                    None => WhereValue::Value(value),
-                    Some(i) => WhereValue::TableColumn(i),
+                    None => RowCheckValue::Static(value),
+                    Some(i) => RowCheckValue::TableColumn(i),
                 }
             },
             2 => {
                 if !splitted_identificator[0].eq(table_name) {
-                    WhereValue::Value(value)
+                    RowCheckValue::Static(value)
                 } else {
                     let index = column_names.iter()
                         .position(|table_column_name| table_column_name.eq(splitted_identificator[1]));
                     match index {
-                        None => WhereValue::Value(value),
-                        Some(i) => WhereValue::TableColumn(i),
+                        None => RowCheckValue::Static(value),
+                        Some(i) => RowCheckValue::TableColumn(i),
                     }
                 }
             },
-            _ => WhereValue::Value(value),
+            _ => RowCheckValue::Static(value),
         }
     }
 }
